Add admin to users table and admin dash for updating databases

+++ ampsite/db/migrate/20150617184819_triodes.rb
@@ -3,11 +3,11 @@ class Triodes < ActiveRecord::Migration
     create_table :triodes do |t|
       t.string :identifier, null: false
       t.string :classification, null: false
-      t.integer :filament_voltage, null: false
-      t.integer :filament_current, null: false
-      t.integer :max_anode_voltage, null: false
-      t.integer :anode_dissipation, null: false
-      t.integer :amplification_factor, null: false
+      t.float :filament_voltage, null: false
+      t.float :filament_current, null: false
+      t.float :max_anode_voltage, null: false
+      t.float :anode_dissipation, null: false
+      t.float :amplification_factor, null: false
 
       t.timestamps
     end
